+++ src/main/java/net/como/client/ComoClient.java
@@ -342,7 +342,7 @@ public class ComoClient {
         this.config = new GeneralConfig();
 
         // Setup the chat command system
-        this.commandHandler = new CommandHandler(isOtherClientLoaded() ? config.alterativeCommandPrefix : config.commandPrefix);
+        this.commandHandler = new CommandHandler(ThirdPartyUtils.isOtherClientLoaded() ? config.alterativeCommandPrefix : config.commandPrefix);
 
         this.registerModules();
 
@@ -381,25 +381,5 @@ public class ComoClient {
         this.modules.put(key, module);
 
         return true;
-    }
-
-    /**
-     * Checks if meteor client is loaded
-     * @return if meteor client is loaded
-     */
-    public static boolean isMeteorLoaded() {
-        return FabricLoader.getInstance().isModLoaded("meteor-client");
-    }
-
-    /**
-     * Checks if coffee client is loaded
-     * @return if coffee client is loaded
-     */
-    public static boolean isCoffeeLoaded() {
-        return FabricLoader.getInstance().isModLoaded("coffee");
-    }
-
-    public static boolean isOtherClientLoaded() {
-        return isMeteorLoaded() || isCoffeeLoaded();
     }
 }
