+++ src/main/java/net/como/client/mixin/botch/TitleScreenLogoMixin.java
@@ -8,6 +8,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import net.como.client.ComoClient;
 import net.como.client.misc.Colour;
 import net.como.client.utils.RenderUtils;
+import net.como.client.utils.ThirdPartyUtils;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.font.TextRenderer;
 import net.minecraft.client.gui.screen.Screen;
@@ -35,7 +36,7 @@ public class TitleScreenLogoMixin extends Screen {
 
         int padding = 2;
         int x = client.getWindow().getScaledWidth() - tr.getWidth(text) - padding;
-        int y = ComoClient.isMeteorLoaded() ? 12 + padding : padding;
+        int y = ThirdPartyUtils.isMeteorLoaded() ? 12 + padding : padding;
 
         tr.drawWithShadow(matrices, Text.of(text), x, y, Colour.GREEN.toARGB());
     }
